task 1

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -8,7 +8,7 @@
 
 size_t list_len(const list_t *h)
 {
-long unsigned int count = 0;
+size_t count = 0;
 
 while (h)
 {
